Some cleaning of S2S. So more needed
