--- scripts/street2shop_categories.json
@@ -1,1 +1,0 @@
-{"bags": 0, "belts": 1, "dresses": 2, "eyewear": 3, "footwear": 4, "hats": 5, "leggings": 6, "outerwear": 7, "pants": 8, "skirts": 9, "tops": 10}
